// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const BookingStatus = {
  PENDING_TO_PAY: 'PENDING_TO_PAY',
  PAID: 'PAID',
  PAYMENT_FAILED: 'PAYMENT_FAILED',
  CANCELLED: 'CANCELLED',
  CHECKED_IN: 'CHECKED_IN',
  CHECKED_OUT: 'CHECKED_OUT'
} as const

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus]


export const HotelType = {
  HOTEL: 'HOTEL',
  MOTEL: 'MOTEL',
  RESORT: 'RESORT',
  APARTMENT: 'APARTMENT',
  HOSTEL: 'HOSTEL'
} as const

export type HotelType = (typeof HotelType)[keyof typeof HotelType]


export const HotelStatus = {
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  REJECTED: 'REJECTED',
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  SUSPENDED: 'SUSPENDED'
} as const

export type HotelStatus = (typeof HotelStatus)[keyof typeof HotelStatus]


export const FacilityType = {
  HOTEL_SERVICES: 'HOTEL_SERVICES',
  FOOD_AND_DRINK: 'FOOD_AND_DRINK',
  PUBLIC: 'PUBLIC',
  IN_ROOM: 'IN_ROOM',
  ACCESSIBILITY: 'ACCESSIBILITY',
  OTHER: 'OTHER'
} as const

export type FacilityType = (typeof FacilityType)[keyof typeof FacilityType]


export const RoomStatus = {
  ACTIVE: 'ACTIVE',
  MAINTENANCE: 'MAINTENANCE',
  OUT_OF_SERVICE: 'OUT_OF_SERVICE'
} as const

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus]


export const BedType = {
  SINGLE: 'SINGLE',
  DOUBLE: 'DOUBLE',
  QUEEN: 'QUEEN',
  KING: 'KING',
  TWIN: 'TWIN'
} as const

export type BedType = (typeof BedType)[keyof typeof BedType]


export const UserRole = {
  ADMIN: 'ADMIN',
  HOTEL_OWNER: 'HOTEL_OWNER',
  USER: 'USER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  PENDING: 'PENDING',
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  DELETED: 'DELETED',
  HOTEL_OWNER_FILLING_INFORMATION: 'HOTEL_OWNER_FILLING_INFORMATION'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const VerificationType = {
  REGISTRATION: 'REGISTRATION',
  PASSWORD_RESET: 'PASSWORD_RESET'
} as const

export type VerificationType = (typeof VerificationType)[keyof typeof VerificationType]